apps/agent/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==14.4.0
black==25.1.0
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
colorama==0.4.6
distro==1.9.0
docstring_parser==0.16
eval_type_backport==0.2.2
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
livekit==1.0.11
livekit-agents==1.1.4
livekit-api==1.0.3
livekit-plugins-noise-cancellation==0.2.4
livekit-plugins-openai==1.1.4
livekit-protocol==1.0.4
multidict==6.6.2
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.1
openai==1.93.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
sniffio==1.3.1
sounddevice==0.5.2
tqdm==4.67.1
types-protobuf==4.25.0.20240417
typing-inspection==0.4.1
typing_extensions==4.14.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
